+++ week8/biker.java
@@ -1,23 +1,23 @@
 package week8;
 public class biker{
-    static int Speed;
+    static double Speed;
     final private static int fixupanddown = 5;
     public biker(){
         
     }
-    public void Speeder(int Speeder){
+    public void Speeder(double Speeder){
        Speed = Speeder;
     }
-    public int Speed1(){
+    public double Speed1(){
         return Speed;
     }
-    public int Speed2(){
+    public double Speed2(){
         return fixupanddown;
     }
     public void resultSpeed(){
         System.out.println("Speed is "+Speed1()+" km/h");
     }
-    public boolean checkingSpeed(int receivespeed){
+    public boolean checkingSpeed(double receivespeed){
         if(receivespeed > 99){
             System.out.println("the speed is can't more then 99 !!!!!");
             return true;
